--- app/src/main/java/com/codesoom/assignment/domain/policy/PolicyStrategy.java
@@ -1,5 +1,0 @@
-package com.codesoom.assignment.domain.policy;
-
-public interface PolicyStrategy {
-    boolean isSatisfiedBy();
-}
