+++ app/src/main/java/com/codesoom/assignment/domain/validator/Validator.java
@@ -1,0 +1,5 @@
+package com.codesoom.assignment.domain.validator;
+
+public interface Validator {
+    boolean isSatisfiedBy();
+}
